Refactor: Restructure files for the website module
