+++ app/[locale]/(website)/books/[id]/page.tsx
