+++ app/[locale]/(website)/components/Footer.tsx
