+++ app/[locale]/(website)/components/Header.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import React from "react";
 import Logo from "./Logo";
 import Link from "next/link";
