+++ app/[locale]/(website)/components/Logo.tsx
