+++ app/[locale]/(website)/layout.tsx
@@ -6,8 +6,7 @@ import {Locale, routing} from '@/i18n/routing';
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "../globals.css";
-import Header from "../components/Header";
-import Footer from "../components/Footer";
+import LayoutWrapper from './LayoutWrapper';
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -46,9 +45,9 @@ export default async function LocaleLayout({
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
         <NextIntlClientProvider messages={messages}>
-          <Header />
-          {children}
-          <Footer />
+          <LayoutWrapper>
+            {children}
+          </LayoutWrapper>
         </NextIntlClientProvider>
       </body>
     </html>
