+++ package.json
@@ -9,21 +9,20 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@formatjs/intl-localematcher": "^0.5.9",
+    "@heroicons/react": "^2.2.0",
     "@stripe/react-stripe-js": "^3.1.1",
     "@stripe/stripe-js": "^5.4.0",
-    "react-icons": "^5.4.0",
-    "stripe": "^17.5.0",
-    "@formatjs/intl-localematcher": "^0.5.9",
-    "negotiator": "^1.0.0",
-    "@heroicons/react": "^2.2.0",
     "autoprefixer": "^10.4.20",
     "axios": "^1.7.9",
     "date-fns": "^4.1.0",
+    "negotiator": "^1.0.0",
     "next": "15.1.0",
     "next-intl": "^3.26.3",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "react-icons": "^5.4.0"
+    "react-icons": "^5.4.0",
+    "stripe": "^17.5.0"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
